requirements.txt
prefect
httpx
duckdb
pandas
pyyaml
python-dotenv
matplotlib
seaborn
plotly
tqdm
jupyter
